0x v2 in subdirectory
