+++ test/e2e/zero-ex/index.ts
@@ -1,0 +1,9 @@
+import { BigNumberish, BytesLike } from "ethers";
+
+export interface SimpleOrder {
+  takerAddress: string;
+  makerAssetAmount: BigNumberish;
+  takerAssetAmount: BigNumberish;
+  makerAssetAddress: BytesLike;
+  takerAssetAddress: BytesLike;
+}
